chore(duktape_ee): move into helper header file
